Crate breez_sdk_core

Breez SDK

The Breez SDK enables mobile developers to integrate Lightning and bitcoin payments into their apps with a very shallow learning curve. The use cases are endless – from social apps that want to integrate tipping between users to content-creation apps interested in adding bitcoin monetization. Crucially, this SDK is an end-to-end, non-custodial, drop-in solution powered by Greenlight, a built-in LSP, on-chain interoperability, third-party fiat on-ramps, and other services users and operators need.

The Breez SDK provides the following services:

Sending payments (via various protocols such as: bolt11, keysend, lnurl-pay, lightning address, etc.)

Receiving payments (via various protocols such as: bolt11, lnurl-withdraw, etc.)

Fetching node status (e.g. balance, max allow to pay, max allow to receive, on-chain balance, etc.)

Connecting to a new or existing node.

Getting Started

The first step is to initialize the SDK and start the node:

let mnemonic = Mnemonic::new(Words12, English);
let seed = Seed::new(&mnemonic, "");

let creds = BreezServices::register_node(Network::Bitcoin, seed.as_bytes().to_vec()).await?;
let sdk = BreezServices::init_services(
        BreezServices::default_config(EnvironmentType::Production),
        seed.to_vec(),
        creds.clone(),
        Box::new(AppEventListener {}),
    )
    .await?;

BreezServices::start(rt(), &sdk).await?;

On initializing the SDK, the caller gets its GreenlightCredentials. These are used to interact with the Greenlight LN node running in the cloud. Together with the BIP39 mnemonic, these can be used to restore access to the same cloud node, either in the same app (backup / restore) or in another app using the SDK.

We can now receive payments

let invoice = sdk.receive_payment(3000, "Invoice for 3000 sats".into()).await?;

or make payments

let bolt11 = "..."; // LN invoice
sdk.send_payment(bolt11.into(), Some(3000)).await?;

At any point we can fetch our balance from the Greenlight node

if let Some(node_state) = sdk.node_info()? {
    let balance_ln = node_state.channels_balance_msat;
    let balance_onchain = node_state.onchain_balance_msat;
}

or fetch other useful infos, like the current mempool RecommendedFees

let fees: RecommendedFees = sdk.recommended_fees().await?;

These different types of operations are described below in more detail.

A. Initializing the SDK

There are two steps necessary to initialize the SDK:

BreezServices::init_services to setup the Breez SDK services

BreezServices::start to start the Greenlight node and all needed Breez SDK services

The first step takes the GreenlightCredentials as an argument. There are three ways to get them:

by loading the credentials from local storage, or with

BreezServices::register_node to register a new Greenlight node, or with

BreezServices::recover_node to recover an existing Greenlight node

The first step also takes an implementation of EventListener as an argument, which is used to notify the caller of SDK events.

After initializing the Breez SDK services and starting them, the SDK is ready to be used.

B. Sending and receiving Lightning payments

Supported BOLT11 operations are

BreezServices::receive_payment to create an invoice

BreezServices::send_payment to pay an invoice

BreezServices::send_spontaneous_payment for keysend payments

C. Receiving an on-chain transaction (swap-in)

BreezServices::receive_onchain

BreezServices::list_refundables to get a list of swaps

BreezServices::refund to broadcast a transaction for failed or expired swaps

D. Using LNURL

parse the LNURL endpoint URL to get the workflow parameters.

After getting the user input or confirmation, complete the workflow with BreezServices::lnurl_pay or BreezServices::lnurl_withdraw.

E. Supporting fiat currencies

BreezServices::list_fiat_currencies to get the supported fiat currencies

BreezServices::fetch_fiat_rates to get the current exchange rates

BreezServices::recommended_fees for the recommended mempool fees

F. Connecting to an LSP

BreezServices::list_lsps to get a list of available LSPs

BreezServices::connect_lsp to connect to a chosen LSP

BreezServices::lsp_info to get LspInformation on the currently selected LSP

G. Utilities

Use parse to parse generic input. The input can come from the user, from a clicked link or from a QR code. The resulting InputType will tell you what the input is and how to treat it, as well as present relevant payload data in a structured form.

Bindings

Dart

Swift

Kotlin

React-Native: coming soon

Modules

binding

Bindings for the Dart integration

Structs

BitcoinAddressData

Wrapped in a BitcoinAddress, this is the result of parse when given a plain or BIP-21 BTC address.

BreezServices

BreezServices is a facade and the single entry point for the SDK.

Channel

Lightning channel

ClosedChannelPaymentDetails

Represents the funds that were on the user side of the channel at the time it was closed.

Config

Configuration for the Breez Services

CurrencyInfo

Details about a supported currency in the fiat rate feed

FiatCurrency

Wrapper around the CurrencyInfo of a fiat currency

GreenlightCredentials

Client-specific credentials to connect to and manage a Greenlight node in the cloud

InvoicePaidDetails

Details of an invoice that has been paid, included as payload in an emitted BreezEvent

LNInvoice

Wrapper for a BOLT11 LN invoice

LnPaymentDetails

Details of a LN payment, as included in a Payment

LnUrlAuthRequestData

Wrapped in a LnUrlAuth, this is the result of parse when given a LNURL-auth endpoint.

LnUrlErrorData

Wrapped in a LnUrlError, this represents a LNURL-endpoint error.

LnUrlPayRequestData

Wrapped in a LnUrlPay, this is the result of parse when given a LNURL-pay endpoint.

LnUrlWithdrawRequestData

Wrapped in a LnUrlWithdraw, this is the result of parse when given a LNURL-withdraw endpoint.

LocaleOverrides

Locale-specific settings for the representation of a currency

LocalizedName

Localized name of a currency

LogEntry

Internal SDK log entry

LspInformation

Details of supported LSP

MetadataItem

Key-value pair in the LnUrlPayRequestData, as returned by the LNURL-pay endpoint

NodeState

The node state of a Greenlight LN node running in the cloud

Payment

Represents a payment, including its PaymentType and PaymentDetails.

Rate

Denominator in an exchange rate

RecommendedFees

Wrapper containing the result of the recommended fees query, in sat/vByte, based on mempool.space data

RouteHint

A route hint for a LN payment

RouteHintHop

Details of a specific hop in a larger route hint

Swap

Summary of an ongoing swap

SwapInfo

Represents the details of an on-going swap.

Symbol

Settings for the symbol representation of a currency

SyncResponse

Internal response to a NodeAPI::pull_changed call

UnspentTransactionOutput

UTXO known to the LN node

Enums

BreezEvent

Event emitted by the SDK. To listen for and react to these events, use an EventListener when initializing the BreezServices.

ChannelState

State of a Lightning channel

EnvironmentType

Indicates the different kinds of supported environments for crate::BreezServices.

FeeratePreset

Different types of supported feerates

InputType

Different kinds of inputs supported by parse, including any relevant details extracted from the input

LnUrlPayResult

Contains the result of the entire LNURL-pay interaction, as reported by the LNURL endpoint.

LnUrlRequestData

Generic struct containing the possible LNURL payloads returned when contacting a LNURL endpoint

LnUrlWithdrawCallbackStatus

Contains the result of the entire LNURL-withdraw interaction, as reported by the LNURL endpoint.

Network

The different supported bitcoin networks

PaymentDetails

Wrapper for the different types of payments

PaymentType

Different types of supported payments

PaymentTypeFilter

Different types of supported filters which can be applied when retrieving the transaction list

SwapStatus

The status of a swap

Traits

EventListener

Trait that can be used to react to various BreezEvents emitted by the SDK.

FiatAPI

Trait covering fiat-related functionality

LspAPI

Trait covering LSP-related functionality

NodeAPI

Trait covering functions affecting the LN node

SwapperAPI

Trait covering functionality involving swaps

Functions

mnemonic_to_seed

Attempts to convert the phrase to a mnemonic, then to a seed.

parse

Parses generic user input, typically pasted from clipboard or scanned from a QR.

parse_invoice

Parse a BOLT11 payment request and return a structure contains the parsed fields.