Enum breez_sdk_core::LnUrlWithdrawCallbackStatus

pub enum LnUrlWithdrawCallbackStatus {
    Ok,
    ErrorStatus {
        data: LnUrlErrorData,
    },
}

Contains the result of the entire LNURL-withdraw interaction, as reported by the LNURL endpoint.

Ok indicates the interaction with the endpoint was valid, and the endpoint started to pay the invoice asynchronously.

Error indicates a generic issue the LNURL endpoint encountered, including a freetext description of the reason.

Both cases are described in LUD-03: https://github.com/lnurl/luds/blob/luds/03.md

Variants

Ok

On-wire format is: {"status": "OK"}

ErrorStatus

Fields

data: LnUrlErrorData

On-wire format is: {"status": "ERROR", "reason": "error details..."}