Struct breez_sdk_core::BreezServices

pub struct BreezServices { /* private fields */ }

BreezServices is a facade and the single entry point for the SDK.

Implementations

impl BreezServices

pub async fn register_node(
network: Network,
seed: Vec<u8>
) -> Result<GreenlightCredentials>

Create a new node for the given network, from the given seed

pub async fn recover_node(
network: Network,
seed: Vec<u8>
) -> Result<GreenlightCredentials>

Try to recover a previously created node

pub async fn init_services(
config: Config,
seed: Vec<u8>,
creds: GreenlightCredentials,
event_listener: Box<dyn EventListener>
) -> Result<Arc<BreezServices>>

Create and initialize the node services instance

pub async fn start(
runtime: &Runtime,
breez_services: &Arc<BreezServices>
) -> Result<()>

Starts the BreezServices background threads for this instance.

It should be called once right after creating BreezServices, since it is essential for the communicating with the node.

It should be called only once when the app is started, regardless whether the app is sent to background and back.

pub async fn stop(&self) -> Result<()>

Trigger the stopping of BreezServices background threads for this instance.

pub async fn send_payment(
&self,
bolt11: String,
amount_sats: Option<u64>
) -> Result<()>

Pay a bolt11 invoice

If the invoice doesn’t specify an amount, the amount is taken from the amount_sats arg.

Arguments

bolt11 - The bolt11 invoice

amount_sats - The amount to pay in satoshis

pub async fn send_spontaneous_payment(
&self,
node_id: String,
amount_sats: u64
) -> Result<()>

Pay directly to a node id using keysend

Arguments

node_id - The destination node_id

amount_sats - The amount to pay in satoshis

pub async fn lnurl_pay(
&self,
user_amount_sat: u64,
comment: Option<String>,
req_data: LnUrlPayRequestData
) -> Result<LnUrlPayResult>

Second step of LNURL-pay. The first step is parse(), which also validates the LNURL destination and generates the LnUrlPayRequestData payload needed here.

This call will validate the given user_amount_sat and comment against the parameters of the LNURL endpoint (req_data). If they match the endpoint requirements, the LNURL payment is made.

pub async fn lnurl_withdraw(
&self,
req_data: LnUrlWithdrawRequestData,
amount_sats: u64,
description: Option<String>
) -> Result<LnUrlWithdrawCallbackStatus>

Second step of LNURL-withdraw. The first step is parse(), which also validates the LNURL destination and generates the LnUrlWithdrawRequestData payload needed here.

This call will validate the given amount_sats against the parameters of the LNURL endpoint (req_data). If they match the endpoint requirements, the LNURL withdraw request is made. A successful result here means the endpoint started the payment.

pub async fn receive_payment(
&self,
amount_sats: u64,
description: String
) -> Result<LNInvoice>

Creates an bolt11 payment request. This also works when the node doesn’t have any channels and need inbound liquidity. In such case when the invoice is paid a new zero-conf channel will be open by the LSP, providing inbound liquidity and the payment will be routed via this new channel.

Arguments

description - The bolt11 payment request description

amount_sats - The amount to receive in satoshis

pub fn node_info(&self) -> Result<Option<NodeState>>

Retrieve the node state from the persistent storage

pub async fn list_payments(
&self,
filter: PaymentTypeFilter,
from_timestamp: Option<i64>,
to_timestamp: Option<i64>
) -> Result<Vec<Payment>>

List payments matching the given filters, as retrieved from persistent storage

pub async fn sweep(
&self,
to_address: String,
fee_rate_sats_per_byte: u64
) -> Result<()>

Sweep on-chain funds to the specified on-chain address, with the given feerate

pub async fn fetch_fiat_rates(&self) -> Result<Vec<Rate>>

Fetch live rates of fiat currencies

pub async fn list_fiat_currencies(&self) -> Result<Vec<FiatCurrency>>

List all supported fiat currencies for which there is a known exchange rate.

pub async fn list_lsps(&self) -> Result<Vec<LspInformation>>

List available LSPs that can be selected by the user

pub async fn connect_lsp(&self, lsp_id: String) -> Result<()>

Select the LSP to be used and provide inbound liquidity

pub async fn lsp_id(&self) -> Result<Option<String>>

Get the current LSP’s ID

pub async fn fetch_lsp_info(&self, id: String) -> Result<Option<LspInformation>>

Convenience method to look up LspInformation for a given LSP ID

pub async fn close_lsp_channels(&self) -> Result<()>

Close all channels with the current LSP.

Should be called when the user wants to close all the channels.

pub async fn receive_onchain(&self) -> Result<SwapInfo>

Onchain receive swap API

Create a SwapInfo that represents the details of an on-going swap.

See SwapInfo for details.

pub async fn list_refundables(&self) -> Result<Vec<SwapInfo>>

List the full swap history. Includes swaps in all states: expired, refunded and active.

pub async fn refund(
&self,
swap_address: String,
to_address: String,
sat_per_vbyte: u32
) -> Result<String>

Construct and broadcast a refund transaction for a failed/expired swap

pub async fn execute_dev_command(&self, command: String) -> Result<String>

Execute a command directly on the NodeAPI interface. Mainly used to debugging.

pub async fn lsp_info(&self) -> Result<LspInformation>

Convenience method to look up LSP info based on current LSP ID

pub async fn recommended_fees(&self) -> Result<RecommendedFees>

Get the recommended fees for onchain transactions

pub fn default_config(env_type: EnvironmentType) -> Config

Get the full default config for a specific environment type