Module breez_sdk_core::binding

Bindings for the Dart integration

Functions

breez_events_stream

breez_log_stream

close_lsp_channels

See BreezServices::close_lsp_channels

connect_lsp

See BreezServices::connect_lsp

default_config

See BreezServices::default_config

execute_command

See BreezServices::execute_dev_command

fetch_fiat_rates

See BreezServices::fetch_fiat_rates

fetch_lsp_info

See BreezServices::fetch_lsp_info

in_progress_swap

See BreezServices::in_progress_swap

init_services

init_services initialized the global NodeService, schedule the node to run in the cloud and run the signer. This must be called in order to start communicate with the node

list_fiat_currencies

See BreezServices::list_fiat_currencies

list_lsps

See BreezServices::list_lsps

list_payments

See BreezServices::list_payments

list_refundables

See BreezServices::list_refundables

lnurl_pay

See BreezServices::lnurl_pay

lnurl_withdraw

See BreezServices::lnurl_withdraw

lsp_id

See BreezServices::lsp_id

mnemonic_to_seed

See breez_services::mnemonic_to_seed

node_info

See BreezServices::node_info

parse

parse_invoice

receive_onchain

See BreezServices::receive_onchain

receive_payment

See BreezServices::receive_payment

recommended_fees

See BreezServices::recommended_fees

recover_node

Recover an existing node from the cloud and return credentials to interact with it

refund

See BreezServices::refund

register_node

Register a new node in the cloud and return credentials to interact with it

send_payment

See BreezServices::send_payment

send_spontaneous_payment

See BreezServices::send_spontaneous_payment

start_node

See BreezServices::start

stop_node

Cleanup node resources and stop the signer.

sweep

See BreezServices::sweep

sync_node